Improvements to the shading calculation and reducing to a single point light while performance is improved.

+++ wGl/assets/shaders/normaldepth-fs.c
@@ -8,7 +8,7 @@ varying vec2 vKdMapCoord;
 uniform sampler2D uMapKd;
 uniform vec2 uMapKdScale;
 
-varying mediump vec4 vNormal;
+varying highp vec4 vNormal;
 varying highp vec4 vPosition;
 varying highp vec4 vpPosition;
